requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
Django==5.1.7
django-crispy-forms==2.3
django-environ==0.12.0
et_xmlfile==2.0.0
openpyxl==3.1.5
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
whitenoise==6.9.0
